Upstream version 9.21

# gpg: Signature made Tue 20 Jan 2015 08:48:01 AM EST using RSA key ID 4A11C97A
# gpg: Good signature from "Ryan Kavanagh <rak@debian.org>"
# gpg:                 aka "Ryan Kavanagh <ryanakca@gmail.com>"
# gpg:                 aka "Ryan Kavanagh <ryanakca@kubuntu.org>"
# gpg:                 aka "Ryan Kavanagh <ryanakca@ubuntu.com>"
# gpg:                 aka "Ryan Kavanagh (Queen's University email alias) <ryan.kavanagh@queensu.ca>"
# gpg:                 aka "Ryan Kavanagh (Queen's University email address) <9rak@queensu.ca>"
# gpg:                 aka "Replacement key for 4EBE56CB0EFECBED2A928D4F7BD15207E95EDDC9"
# gpg:                 aka "Ryan Kavanagh <ryan@cs.queensu.ca>"
# gpg:                 aka "Ryan Kavanagh <ryank@cs.mcgill.ca>"
# gpg:                 aka "Ryan Kavanagh <ryan.kavanagh@mail.mcgill.ca>"
# gpg:                 aka "Ryan Kavanagh <rakavan@mit.edu>"
# gpg:                 aka "Ryan Kavanagh <rak@ryanak.ca>"
# gpg:                 aka "Ryan Kavanagh <ryan.kavanagh@ens-lyon.fr>"

+++ src/rxvtimg.C
@@ -8,7 +8,7 @@
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
+ * the Free Software Foundation; either version 3 of the License, or
  * (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
@@ -190,11 +190,12 @@ namespace
     // CreateSolidFill creates a very very very weird picture
     void mask (const rgba &c)
     {
+      // the casts are needed in C++11 (see 8.5.1)
       XRenderColor rc = {
-        c.r * c.a / 65535,
-        c.g * c.a / 65535,
-        c.b * c.a / 65535,
-              c.a
+        (unsigned short)(c.r * c.a / 65535),
+        (unsigned short)(c.g * c.a / 65535),
+        (unsigned short)(c.b * c.a / 65535),
+        c.a
       };
       msk = XRenderCreateSolidFill (dpy, &rc);
       ecb_assume (msk);
@@ -203,10 +204,10 @@ namespace
     void fill (const rgba &c)
     {
       XRenderColor rc = {
-        c.r * c.a / 65535,
-        c.g * c.a / 65535,
-        c.b * c.a / 65535,
-              c.a
+        (unsigned short)(c.r * c.a / 65535),
+        (unsigned short)(c.g * c.a / 65535),
+        (unsigned short)(c.b * c.a / 65535),
+        c.a
       };
 
       XRenderFillRectangle (dpy, PictOpSrc, msk, &rc, 0, 0, 1, 1);
